+++ src/core/cameras/Camera.hpp
@@ -105,9 +105,9 @@ public:
     inline Vec3f getLinear(int x, int y) const
     {
         int idx = x + y*_res.x();
-        Vec3d result = _pixels[idx]/max(_weights[idx], 1u);
+        Vec3d result = _pixels[idx]/double(max(_weights[idx], 1u));
         if (!_splats.empty())
-            result += _splats[idx]/max(_splatWeight, 1ull);
+            result += _splats[idx]/double(max(_splatWeight, 1ull));
         return Vec3f(result);
     }
