+++ src/core/io/ZipReader.cpp
@@ -8,11 +8,11 @@ namespace Tungsten {
 
 static size_t minizFileReadFunc(void *userPtr, mz_uint64 file_ofs, void *pBuf, size_t n)
 {
-	std::istream &in = *static_cast<std::istream *>(userPtr);
+    std::istream &in = *static_cast<std::istream *>(userPtr);
     if (std::streampos(file_ofs) != in.tellg())
         in.seekg(file_ofs);
     in.read(reinterpret_cast<char *>(pBuf), n);
-    return in.gcount();
+    return size_t(in.gcount());
 }
 
 ZipReader::ZipReader(const Path &p)
@@ -23,7 +23,7 @@ ZipReader::ZipReader(const Path &p)
         FAIL("Failed to construct ZipReader: Input stream is invalid");
     uint64 size = FileUtils::fileSize(p);
     if (size == 0)
-    	FAIL("Failed to construct ZipReader: Error reading file size");
+        FAIL("Failed to construct ZipReader: Error reading file size");
 
     std::memset(&_archive, 0, sizeof(mz_zip_archive));
     _archive.m_pRead = &minizFileReadFunc;
@@ -33,17 +33,17 @@ ZipReader::ZipReader(const Path &p)
 
     int count = mz_zip_reader_get_num_files(&_archive);
     for (int i = 0; i < count; ++i) {
-    	mz_zip_archive_file_stat stat;
-    	mz_zip_reader_file_stat(&_archive, i, &stat);
+        mz_zip_archive_file_stat stat;
+        mz_zip_reader_file_stat(&_archive, i, &stat);
 
-    	ZipEntry entry;
-    	entry.fullPath = Path(stat.m_filename);
-    	entry.name = entry.fullPath.fileName();
-    	entry.size = stat.m_uncomp_size;
-    	entry.isDirectory = !entry.fullPath.empty() && entry.fullPath.asString().back() == '/';
-    	entry.archiveIndex = i;
+        ZipEntry entry;
+        entry.fullPath = Path(stat.m_filename);
+        entry.name = entry.fullPath.fileName();
+        entry.size = uint32(stat.m_uncomp_size);
+        entry.isDirectory = !entry.fullPath.empty() && entry.fullPath.asString().back() == '/';
+        entry.archiveIndex = i;
 
-    	addPath(Path(stat.m_filename).stripSeparator(), std::move(entry));
+        addPath(Path(stat.m_filename).stripSeparator(), std::move(entry));
     }
 }
 
@@ -51,7 +51,7 @@ int ZipReader::addPath(const Path &p, ZipEntry entry)
 {
     auto iter = _pathToEntry.find(p);
     if (iter == _pathToEntry.end()) {
-        int index = _entries.size();
+        int index = int(_entries.size());
 
         _pathToEntry.insert(std::make_pair(p, index));
         _entries.emplace_back(std::move(entry));
